spliting tests because it was getting huge
